requirements.txt
aspose-slides
aspose-words
argparse
colorama 
aspose-cells